update minor changes to code
