--- src/pages/home copy/Menu/Menu.css
@@ -1,9 +1,0 @@
-.hidden {
-  display: none;
-}
-
-@media only screen and (min-width: 935px) {
-  #menuIntro {
-    max-width: 935px;
-  }
-}
