--- src/pages/home copy/Menu/PizzaItemLayout/PizzaItemLayout.css
@@ -1,13 +1,0 @@
-#pizzaItem {
-  width: 26em;
-}
-
-#pizzaItem img {
-  width: 13em;
-  object-fit: contain;
-}
-
-#pizzaaItem #right {
-  max-width: calc(100% - 170px);
-  flex-basis: 0;
-}
