+++ src/pages/home/Promo/Promo.css
@@ -1,0 +1,7 @@
+#promoBox {
+  border: 2px dashed #fff;
+}
+
+#promoBox {
+  background-color: rgba(200, 18, 47);
+}
